+++ src/main/java/com/challenger/fridge/domain/Notification.java
@@ -2,11 +2,8 @@ package com.challenger.fridge.domain;
 
 import static jakarta.persistence.FetchType.*;
 
-import com.challenger.fridge.common.StorageStatus;
-import com.challenger.fridge.domain.box.StorageBox;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
@@ -14,7 +11,6 @@ import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 
 import java.time.LocalDate;
-import java.util.List;
 
 import lombok.AccessLevel;
 import lombok.Getter;
@@ -23,16 +19,16 @@ import lombok.NoArgsConstructor;
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Notice {
+public class Notification {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "notice_id")
+    @Column(name = "notification_id")
     private Long id;
 
     private String message;
 
-    private boolean isRead;
+    private Boolean isRead;
 
     private LocalDate createdDate;
 
@@ -40,22 +36,20 @@ public class Notice {
     @JoinColumn(name = "member_id")
     private Member member;
 
-
-    public Notice(String message, boolean isRead, LocalDate createdDate, Member member) {
+    protected Notification(String message, boolean isRead, LocalDate createdDate, Member member) {
         this.message = message;
         this.isRead = isRead;
         this.createdDate = createdDate;
         this.member = member;
     }
 
-    public static Notice createNotice(Member member, String message) {
-        Notice notice = new Notice(
+    public static Notification createNotice(Member member, String message) {
+        return new Notification(
                 message,
                 false,
                 LocalDate.now(),
                 member
         );
-        return notice;
     }
 
 }
