--- src/main/java/com/challenger/fridge/dto/notice/DeviceTokenRequest.java
@@ -1,10 +1,0 @@
-package com.challenger.fridge.dto.notice;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class DeviceTokenRequest {
-    private String targetToken;
-}
