--- src/main/java/com/challenger/fridge/dto/notice/FcmMessage.java
@@ -1,30 +1,0 @@
-package com.challenger.fridge.dto.notice;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-
-@Builder
-@AllArgsConstructor
-@Getter
-public class FcmMessage {
-    private boolean validateOnly;
-    private Message message;
-
-    @Builder
-    @AllArgsConstructor
-    @Getter
-    public static class Message {
-        private Notification notification;
-        private String token;
-    }
-
-    @Builder
-    @AllArgsConstructor
-    @Getter
-    public static class Notification {
-        private String title;
-        private String body;
-        private String image;
-    }
-}
