--- src/main/java/com/challenger/fridge/repository/NoticeRepository.java
@@ -1,8 +1,0 @@
-package com.challenger.fridge.repository;
-
-import com.challenger.fridge.domain.Notice;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface NoticeRepository extends JpaRepository<Notice, Long> {
-
-}
